feat: add and refactor VectorStore iterators/providers for all supported backends; update RowPagingProvider and related classes to use new generic iterator pattern
