+++ src/main/java/org/mule/extension/vectors/internal/store/BaseStoreService.java
@@ -161,51 +161,4 @@ public abstract class BaseStoreService implements VectorStoreService {
       embeddingStore.removeAll();
     }
   }
-
-  public abstract Iterator<Row<?>> getRowIterator();
-
-
-  public static class Row<Embedded> {
-
-    String id;
-    Embedding embedding;
-    Embedded embedded;
-
-    public Row(String id, Embedding embedding) {
-      this(id, embedding, (Embedded)null);
-    }
-
-    public Row(String id, Embedding embedding, Embedded embedded) {
-      this.id = ValidationUtils.ensureNotBlank(id, "id");
-      this.embedding = embedding;
-      this.embedded = embedded;
-    }
-
-    public boolean equals(Object o) {
-      if (this == o) {
-        return true;
-      } else if (o != null && this.getClass() == o.getClass()) {
-        Row<?> that = (Row)o;
-        return Objects.equals(this.id, that.id) && Objects.equals(this.embedding, that.embedding) && Objects.equals(this.embedded, that.embedded);
-      } else {
-        return false;
-      }
-    }
-
-    public int hashCode() {
-      return Objects.hash(new Object[]{this.id, this.embedding, this.embedded});
-    }
-
-    public String getId() {
-      return id;
-    }
-
-    public Embedding getEmbedding() {
-      return embedding;
-    }
-
-    public Embedded getEmbedded() {
-      return embedded;
-    }
-  }
 }
