+++ src/main/java/org/mule/extension/vectors/internal/store/chroma/ChromaStore.java
@@ -1,7 +1,5 @@
 package org.mule.extension.vectors.internal.store.chroma;
 
-import dev.langchain4j.data.document.Metadata;
-import dev.langchain4j.data.embedding.Embedding;
 import dev.langchain4j.data.segment.TextSegment;
 import dev.langchain4j.store.embedding.EmbeddingStore;
 import dev.langchain4j.store.embedding.chroma.ChromaEmbeddingStore;
@@ -23,11 +21,6 @@ import org.slf4j.LoggerFactory;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
@@ -130,105 +123,4 @@ public class ChromaStore extends BaseStoreService {
         return collection.getString("id");
     }
 
-    @Override
-    public Iterator<BaseStoreService.Row<?>> getRowIterator() {
-        try {
-            return new RowIterator();
-        } catch (IOException e) {
-            throw new ModuleException("Error creating Chroma iterator", MuleVectorsErrorType.STORE_SERVICES_FAILURE, e);
-        }
-    }
-
-    public class RowIterator implements Iterator<BaseStoreService.Row<?>> {
-        private final List<String> ids = new ArrayList<>();
-        private final List<JSONObject> metadata = new ArrayList<>();
-        private final List<String> documents = new ArrayList<>();
-        private final List<JSONArray> embeddings = new ArrayList<>();
-        private int totalCount = 0;
-        private int currentPage = 0;
-        private int pageIndex = 0;
-
-        public RowIterator() throws IOException {
-            this.totalCount = getSegmentCount();
-            loadNextPage();
-        }
-
-        private void loadNextPage() throws IOException {
-            if (ids.size() >= totalCount) {
-                return;
-            }
-
-            JSONObject jsonRequest = new JSONObject();
-            jsonRequest.put("limit", queryParams.pageSize());
-            jsonRequest.put("offset", (long) currentPage * queryParams.pageSize());
-
-            JSONArray jsonInclude = new JSONArray();
-            jsonInclude.put(METADATA_DEFAULT_FIELD_NAME);
-            jsonInclude.put(TEXT_DEFAULT_FIELD_NAME);
-            if (queryParams.retrieveEmbeddings()) {
-                jsonInclude.put(VECTOR_DEFAULT_FIELD_NAME);
-            }
-            jsonRequest.put("include", jsonInclude);
-
-            String jsonResponse = getJsonResponse("/api/v1/collections/" + collectionId + "/get", jsonRequest.toString());
-
-            JSONObject responseObject = new JSONObject(jsonResponse);
-
-            // Clear old data
-            ids.clear();
-            metadata.clear();
-            documents.clear();
-            embeddings.clear();
-            pageIndex = 0;
-
-            // Populate with new data
-            responseObject.optJSONArray(ID_DEFAULT_FIELD_NAME).forEach(id -> ids.add(id.toString()));
-            responseObject.optJSONArray(METADATA_DEFAULT_FIELD_NAME).forEach(meta -> metadata.add((JSONObject) meta));
-            responseObject.optJSONArray(TEXT_DEFAULT_FIELD_NAME).forEach(doc -> documents.add(doc.toString()));
-            if (queryParams.retrieveEmbeddings()) {
-                responseObject.optJSONArray(VECTOR_DEFAULT_FIELD_NAME).forEach(emb -> embeddings.add((JSONArray) emb));
-            }
-            currentPage++;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return (((currentPage-1) * queryParams.pageSize()) + pageIndex) < totalCount;
-        }
-
-        @Override
-        public BaseStoreService.Row<?> next() {
-            if (!hasNext()) {
-                throw new NoSuchElementException();
-            }
-
-            if (pageIndex >= ids.size()) {
-                try {
-                    loadNextPage();
-                } catch (IOException e) {
-                    throw new ModuleException("Failed to load next page from Chroma", MuleVectorsErrorType.STORE_SERVICES_FAILURE, e);
-                }
-            }
-
-            String id = ids.get(pageIndex);
-            Map<String, Object> metadataMap = metadata.get(pageIndex).toMap();
-            String text = documents.get(pageIndex);
-            JSONArray vectorArray = queryParams.retrieveEmbeddings() ? embeddings.get(pageIndex) : null;
-
-            Embedding embedding = null;
-            if (vectorArray != null) {
-                float[] vector = new float[vectorArray.length()];
-                for (int i = 0; i < vectorArray.length(); i++) {
-                    vector[i] = (float) vectorArray.getDouble(i);
-                }
-                embedding = new Embedding(vector);
-            }
-
-            TextSegment textSegment = new TextSegment(text, Metadata.from(metadataMap));
-
-            pageIndex++;
-
-            return new BaseStoreService.Row<>(id, embedding, textSegment);
-        }
-    }
 }
