+++ src/main/java/org/mule/extension/vectors/internal/store/pinecone/PineconeStoreServiceProvider.java
@@ -1,25 +1,25 @@
-package org.mule.extension.vectors.internal.store.chroma;
+package org.mule.extension.vectors.internal.store.pinecone;
 
 import org.mule.extension.vectors.internal.config.StoreConfiguration;
-import org.mule.extension.vectors.internal.connection.store.alloydb.AlloyDBStoreConnection;
-import org.mule.extension.vectors.internal.connection.store.chroma.ChromaStoreConnection;
+import org.mule.extension.vectors.internal.connection.store.pgvector.PGVectorStoreConnection;
+import org.mule.extension.vectors.internal.connection.store.pinecone.PineconeStoreConnection;
 import org.mule.extension.vectors.internal.helper.parameter.QueryParameters;
 import org.mule.extension.vectors.internal.service.VectorStoreService;
 import org.mule.extension.vectors.internal.service.VectorStoreServiceProvider;
 import org.mule.extension.vectors.internal.service.VectoreStoreIterator;
 
-public class ChromaStoreServiceProvider implements VectorStoreServiceProvider {
+public class PineconeStoreServiceProvider implements VectorStoreServiceProvider {
 
   private final StoreConfiguration storeConfiguration;
-  private final ChromaStoreConnection chromaStoreConnection;
+  private final PineconeStoreConnection connection;
   private final String storeName;
   private final QueryParameters queryParams;
   private final int dimension;
   private final boolean createStore;
 
-  public ChromaStoreServiceProvider(StoreConfiguration compositeConfiguration, ChromaStoreConnection chromaStoreConnection, String storeName, QueryParameters queryParams, int dimension, boolean createStore) {
+  public PineconeStoreServiceProvider(StoreConfiguration compositeConfiguration, PineconeStoreConnection connection, String storeName, QueryParameters queryParams, int dimension, boolean createStore) {
     this.storeConfiguration = compositeConfiguration;
-    this.chromaStoreConnection = chromaStoreConnection;
+    this.connection = connection;
     this.storeName = storeName;
     this.queryParams = queryParams;
     this.dimension = dimension;
@@ -27,11 +27,12 @@ public class ChromaStoreServiceProvider implements VectorStoreServiceProvider {
   }
   @Override
   public VectorStoreService getService() {
-    return new ChromaStore(storeConfiguration, chromaStoreConnection, storeName, queryParams, dimension, createStore);
+    return new PineconeStore(storeConfiguration, connection, storeName, queryParams, dimension, createStore);
+
   }
 
   @Override
   public VectoreStoreIterator getFileIterator() {
-    return new ChromaStoreIterator(chromaStoreConnection, storeName, queryParams);
+    return new PineconeStoreIterator(connection, storeName, queryParams);
   }
 }
