+++ src/main/java/org/mule/extension/vectors/internal/store/weaviate/WeaviateStore.java
@@ -4,8 +4,7 @@ import dev.langchain4j.data.segment.TextSegment;
 import dev.langchain4j.exception.AuthenticationException;
 import dev.langchain4j.store.embedding.EmbeddingStore;
 import dev.langchain4j.store.embedding.weaviate.WeaviateEmbeddingStore;
-import io.weaviate.client.base.WeaviateErrorMessage;
-import io.weaviate.client.base.http.HttpResponse;
+
 
 import org.mule.extension.vectors.internal.config.StoreConfiguration;
 import org.mule.extension.vectors.internal.connection.store.weaviate.WeaviateStoreConnection;
@@ -15,12 +14,8 @@ import org.mule.extension.vectors.internal.store.BaseStoreService;
 import org.mule.runtime.extension.api.exception.ModuleException;
 
 
-import java.io.IOException;
-import java.net.ConnectException;
-import java.util.Iterator;
-import java.util.NoSuchElementException;
 import java.util.Optional;
-import java.util.stream.Collectors;
+
 
 /**
  * WeaviateStore is a specialized implementation of {@link BaseStoreService} designed to interact with
@@ -66,43 +61,6 @@ public class WeaviateStore extends BaseStoreService {
             throw new ModuleException("Authentication failed: " + errorMessage, MuleVectorsErrorType.AUTHENTICATION, e);
         } catch (Exception e) {
             throw new ModuleException("Failed to build Weaviate embedding store: " + e.getMessage(), MuleVectorsErrorType.STORE_SERVICES_FAILURE, e);
-        }
-    }
-
-    @Override
-    public Iterator<BaseStoreService.Row<?>> getRowIterator() {
-        // TODO: Implement Weaviate RowIterator to allow paginated fetching of all vectors.
-        // This will require using the Weaviate Java client directly to iterate through results,
-        // similar to the PineconeStore implementation.
-        throw new ModuleException("This operation is not yet supported for Weaviate.", MuleVectorsErrorType.STORE_UNSUPPORTED_OPERATION);
-    }
-
-    public class RowIterator implements Iterator<BaseStoreService.Row<?>> {
-
-
-        public RowIterator() {
-            // TODO: Initialize Weaviate client and prepare for iteration.
-        }
-
-        @Override
-        public boolean hasNext() {
-            // TODO: Check if there are more results to fetch.
-            return false;
-        }
-
-        @Override
-        public BaseStoreService.Row<?> next() {
-            if (!hasNext()) {
-                throw new NoSuchElementException();
-            }
-            try {
-                // TODO: Fetch the next batch of results and return the next row.
-                return null;
-
-            } catch (Exception e) {
-                LOGGER.error("Error while fetching next row", e);
-                throw new NoSuchElementException("No more elements available");
-            }
         }
     }
 }
