+++ src/main/java/org/mule/extension/vectors/internal/store/weaviate/WeaviateStoreIterator.java
@@ -1,0 +1,73 @@
+package org.mule.extension.vectors.internal.store.weaviate;
+
+import org.mule.extension.vectors.internal.service.VectoreStoreIterator;
+import org.mule.extension.vectors.internal.store.VectorStoreRow;
+import dev.langchain4j.data.embedding.Embedding;
+import dev.langchain4j.data.segment.TextSegment;
+import dev.langchain4j.data.document.Metadata;
+import org.mule.extension.vectors.internal.error.MuleVectorsErrorType;
+import org.mule.extension.vectors.internal.helper.parameter.QueryParameters;
+import org.mule.runtime.extension.api.exception.ModuleException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.NoSuchElementException;
+
+public class WeaviateStoreIterator<Embedded> implements VectoreStoreIterator<VectorStoreRow<Embedded>> {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(WeaviateStoreIterator.class);
+
+  private final QueryParameters queryParams;
+
+
+  public WeaviateStoreIterator(
+      // WeaviateClient weaviateClient,
+      QueryParameters queryParams
+      // ...other params as needed
+  ) {
+    // this.weaviateClient = weaviateClient;
+    this.queryParams = queryParams;
+    // this.currentBatch = new ArrayList<>();
+    // this.currentIndex = 0;
+    // this.hasMorePages = true;
+    // fetchNextBatch();
+  }
+
+  @Override
+  public boolean hasNext() {
+    // TODO: Implement actual paging logic using Weaviate client
+    // return currentIndex < currentBatch.size() || fetchNextBatch();
+    return false;
+  }
+
+  @Override
+  public VectorStoreRow<Embedded> next() {
+    if (!hasNext()) {
+      throw new NoSuchElementException();
+    }
+    try {
+      // TODO: Fetch the next result from Weaviate and convert to VectorStoreRow<Embedded>
+      // Example (replace with actual logic):
+      // WeaviateResultType result = currentBatch.get(currentIndex++);
+      // String id = ...;
+      // float[] vector = ...;
+      // String text = ...;
+      // Map<String, Object> metadataMap = ...;
+      // @SuppressWarnings("unchecked")
+      // Embedded embedded = (Embedded) new TextSegment(text, Metadata.from(metadataMap));
+      // return new VectorStoreRow<>(id, vector != null ? new Embedding(vector) : null, embedded);
+      return null;
+    } catch (Exception e) {
+      LOGGER.error("Error while fetching next row", e);
+      throw new NoSuchElementException("No more elements available");
+    }
+  }
+
+  // private boolean fetchNextBatch() {
+  //     // TODO: Implement batch fetching logic using Weaviate client
+  //     return false;
+  // }
+}
