+++ src/main/java/org/mule/extension/vectors/internal/store/weaviate/WeaviateStoreServiceProvider.java
@@ -1,25 +1,26 @@
-package org.mule.extension.vectors.internal.store.chroma;
+package org.mule.extension.vectors.internal.store.weaviate;
 
 import org.mule.extension.vectors.internal.config.StoreConfiguration;
-import org.mule.extension.vectors.internal.connection.store.alloydb.AlloyDBStoreConnection;
-import org.mule.extension.vectors.internal.connection.store.chroma.ChromaStoreConnection;
+import org.mule.extension.vectors.internal.connection.store.qdrant.QdrantStoreConnection;
+import org.mule.extension.vectors.internal.connection.store.weaviate.WeaviateStoreConnection;
 import org.mule.extension.vectors.internal.helper.parameter.QueryParameters;
 import org.mule.extension.vectors.internal.service.VectorStoreService;
 import org.mule.extension.vectors.internal.service.VectorStoreServiceProvider;
 import org.mule.extension.vectors.internal.service.VectoreStoreIterator;
 
-public class ChromaStoreServiceProvider implements VectorStoreServiceProvider {
+public class WeaviateStoreServiceProvider implements VectorStoreServiceProvider {
+
 
   private final StoreConfiguration storeConfiguration;
-  private final ChromaStoreConnection chromaStoreConnection;
+  private final WeaviateStoreConnection connection;
   private final String storeName;
   private final QueryParameters queryParams;
   private final int dimension;
   private final boolean createStore;
 
-  public ChromaStoreServiceProvider(StoreConfiguration compositeConfiguration, ChromaStoreConnection chromaStoreConnection, String storeName, QueryParameters queryParams, int dimension, boolean createStore) {
+  public WeaviateStoreServiceProvider(StoreConfiguration compositeConfiguration, WeaviateStoreConnection connection, String storeName, QueryParameters queryParams, int dimension, boolean createStore) {
     this.storeConfiguration = compositeConfiguration;
-    this.chromaStoreConnection = chromaStoreConnection;
+    this.connection = connection;
     this.storeName = storeName;
     this.queryParams = queryParams;
     this.dimension = dimension;
@@ -27,11 +28,12 @@ public class ChromaStoreServiceProvider implements VectorStoreServiceProvider {
   }
   @Override
   public VectorStoreService getService() {
-    return new ChromaStore(storeConfiguration, chromaStoreConnection, storeName, queryParams, dimension, createStore);
+    return new WeaviateStore(storeConfiguration,  connection, storeName, queryParams, dimension, createStore);
+
   }
 
   @Override
   public VectoreStoreIterator getFileIterator() {
-    return new ChromaStoreIterator(chromaStoreConnection, storeName, queryParams);
+    return new WeaviateStoreIterator(queryParams);
   }
 }
